modernize to angular19
Use inject instead of constructor, use @if and @for instead of *ngIf etc.
and rely on signals when displaying data in html. Also use more operators
to improve code understanding
