+++ angular-frontend/src/app/app.component.html
@@ -3,4 +3,4 @@
 </nav>
 
 <router-outlet></router-outlet>
-<app-device-list></app-device-list>
+<!-- <app-device-list></app-device-list> -->
