--- angular-frontend/src/app/device-handling/devicelist.component.html
@@ -1,21 +1,0 @@
-<div>
-    <button (click)="getDevicesList()">Load devices list </button>
-    <div *ngIf="deviceListResponse">
-        <div>
-            <h3>Devices</h3>
-            <ul>
-                <li *ngFor="let device of deviceListResponse.devices">
-                    UUID: {{ device.UUID }}
-                </li>
-            </ul>
-        </div>
-        <div>
-            <h3>Colors</h3>
-            <ul>
-                <li *ngFor="let colorObject of deviceListResponse.colors">
-                    Color - R: {{ colorObject.color.r }}, G: {{ colorObject.color.g }}, B: {{ colorObject.color.b }}
-                </li>
-            </ul>
-        </div>
-    </div>
-</div>
