+++ angular-frontend/src/app/omnai-datasource/devicelist.component.html
@@ -1,0 +1,25 @@
+<button (click)="getDevicesList()">Load devices list </button>
+@if(deviceListResponse(); as list){<div *ngIf="deviceListResponse">
+    <div>
+        <h3>Devices</h3>
+        <ul>
+            @for (device of list.devices; track device.UUID) {<li>
+                UUID: {{ device.UUID }}
+            </li>
+            }
+        </ul>
+    </div>
+    <div>
+        <h3>Colors</h3>
+        <ul>
+            @for (color of list.colors; track $index) {
+            <li>
+                Color - R: {{ color.color.r }},
+                G: {{ color.color.g }},
+                B: {{ color.color.b }}
+            </li>}
+
+        </ul>
+    </div>
+</div>
+}
